+++ BUS/ViewModel/ProductDetailViewModel.cs
@@ -8,8 +8,8 @@ namespace BUS.ViewModel
 
         public string ProductName { get; set; }
         public string SkuId { get; set; }
-        public float ImportPrice { get; set; }
-        public float Price { get; set; }
+        public Int64 ImportPrice { get; set; }
+        public Int64 Price { get; set; }
         public int Quantity { get; set; }
         public string Description { get; set; }
         public ICollection<ImageValueObject>? Images { get; set; }
