+++ DAL/Entities/OrderDetails.cs
@@ -7,7 +7,7 @@ namespace DAL.Entities
         public Guid OrderID { get; set; }
         public Guid VariantID { get; set; }
         public int Quantity { get; set; }
-        public float UnitPrice { get; set; }
+        public Int64 UnitPrice { get; set; }
         public bool IsOrderDetailEnabled { get; set; }
         public Order Orderds { get; set; }
         public ProductVariants ProductVariants { get; set; }
